# backend/CafeCRM/src/main/java/org/example/cafecrm/dto/CreateMenuCategoryRequest.java
package org.example.cafecrm.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record CreateMenuCategoryRequest(
        @NotBlank @Size(max = 50) String name
) {}